+++ package.json
@@ -39,6 +39,7 @@
     "apollo-server": "^2.10.0",
     "envalid": "^6.0.1",
     "jest": "^25.1.0",
+    "reekwire": "^1.0.0",
     "semver": "^7.1.3"
   },
   "devDependencies": {
